requirements.txt
scapy==2.5.0
netifaces==0.11.0
colorama==0.4.6
python-nmap==0.7.1
netaddr==0.8.0
PyQt6==6.6.1
requests==2.31.0
cryptography==42.0.2
python-whois==0.8.0
dnspython==2.5.0
beautifulsoup4==4.12.3
aiohttp==3.9.3
asyncio==3.4.3
pyOpenSSL==24.0.0
nest-asyncio==1.6.0 